Reorganize the job configs into one file.
I also plan to move plugins in here after this goes in.

This will have a few benefits in the long run:
* Adding new job types won't require so much duplication.
* The config is passed to plugins. Currently there is no way for plugins
to configure themselves at runtime. Passing them this config makes it
possible without adding yet more configmaps.
* Separating Config and ConfigAgent makes sense. I was using a JobAgent
in several places that only needed to load the job configs once. With
this change, ConfigAgent mostly just knows how to return a read-only
Config struct.
