+++ prow/BUILD
@@ -25,10 +25,10 @@ filegroup(
         "//prow/cmd/sinker:all-srcs",
         "//prow/cmd/splice:all-srcs",
         "//prow/cmd/tot:all-srcs",
+        "//prow/config:all-srcs",
         "//prow/crier:all-srcs",
         "//prow/github:all-srcs",
         "//prow/jenkins:all-srcs",
-        "//prow/jobs:all-srcs",
         "//prow/kube:all-srcs",
         "//prow/line:all-srcs",
         "//prow/plugins:all-srcs",
