+++ prow/README.md
@@ -20,10 +20,10 @@ Build with `bazel build //prow/...`, test with `bazel test //prow/...`.
 
 You can run `cmd/hook` in a local mode for testing, and hit it with arbitrary
 fake webhooks. To do this, run `./bazel-bin/prow/cmd/hook/hook --local
---presubmits prow/presubmit.yaml --postsubmits prow/postsubmit.yaml
---plugin-config prow/plugins.yaml` in one shell. This will listen on
-`localhost:8888` for webhooks. Send one with `./bazel-bin/prow/cmd/phony/phony
---event issue_comment --payload prow/cmd/phony/examples/test_comment.json`.
+--config prow/config.yaml --plugin-config prow/plugins.yaml` in one shell. This
+will listen on `localhost:8888` for webhooks. Send one with
+`./bazel-bin/prow/cmd/phony/phony --event issue_comment --payload
+prow/cmd/phony/examples/test_comment.json`.
 
 ## How to update the cluster
 
@@ -57,9 +57,9 @@ redeploying the binaries, and will take effect within a minute.
 
 ## How to add new jobs
 
-To add a new job you'll need to add an entry into `presubmit.yaml`. Then run
-`make update-jobs`. This does not require redeploying any binaries, and will
-take effect within a minute.
+To add a new job you'll need to add an entry into `config.yaml`. Then run `make
+update-config`. This does not require redeploying any binaries, and will take
+effect within a minute.
 
 The Jenkins job itself should have no trigger. It will be called with string
 parameters `PULL_NUMBER` and `PULL_BASE_REF` which it can use to checkout the
@@ -70,4 +70,4 @@ appropriate revision. It needs to accept the `buildId` parameter which the
 
 [@k8s-ci-robot](https://github.com/k8s-ci-robot) and its silent counterpart
 [@k8s-bot](https://github.com/k8s-bot) both live here as triggers to GitHub
-messages defined in [presubmit.yaml](presubmit.yaml).
+messages defined in [config.yaml](config.yaml).
