+++ prow/cluster/hook_deployment.yaml
@@ -35,7 +35,7 @@ spec:
       terminationGracePeriodSeconds: 30
       containers:
       - name: hook
-        image: gcr.io/k8s-prow/hook:0.81
+        image: gcr.io/k8s-prow/hook:0.82
         imagePullPolicy: Always
         args:
         - "--github-bot-name=k8s-ci-robot"
@@ -54,8 +54,8 @@ spec:
         - name: oauth
           mountPath: /etc/github
           readOnly: true
-        - name: job-configs
-          mountPath: /etc/jobs
+        - name: config
+          mountPath: /etc/config
           readOnly: true
         - name: plugins
           mountPath: /etc/plugins
@@ -67,9 +67,9 @@ spec:
       - name: oauth
         secret:
           secretName: oauth-token
-      - name: job-configs
+      - name: config
         configMap:
-          name: job-configs
+          name: config
       - name: plugins
         configMap:
           name: plugins
