+++ prow/cluster/splice_deployment.yaml
@@ -15,10 +15,10 @@ spec:
         role: prow
       containers:
       - name: splice
-        image: gcr.io/k8s-prow/splice:0.12
+        image: gcr.io/k8s-prow/splice:0.13
         volumeMounts:
-        - name: job-configs
-          mountPath: /etc/jobs
+        - name: config
+          mountPath: /etc/config
           readOnly: true
         args:
         - -log-json
@@ -28,6 +28,6 @@ spec:
         - name: DRY_RUN
           value: "false"
       volumes:
-      - name: job-configs
+      - name: config
         configMap:
-          name: job-configs
+          name: config
