+++ prow/cmd/hook/BUILD
@@ -41,8 +41,8 @@ go_library(
     ],
     tags = ["automanaged"],
     deps = [
+        "//prow/config:go_default_library",
         "//prow/github:go_default_library",
-        "//prow/jobs:go_default_library",
         "//prow/kube:go_default_library",
         "//prow/plugins:go_default_library",
         "//prow/plugins/cla:go_default_library",
