+++ prow/cmd/line/BUILD
@@ -24,7 +24,7 @@ go_test(
     library = ":go_default_library",
     tags = ["automanaged"],
     deps = [
-        "//prow/jobs:go_default_library",
+        "//prow/config:go_default_library",
         "//prow/line:go_default_library",
     ],
 )
@@ -37,10 +37,10 @@ go_library(
     ],
     tags = ["automanaged"],
     deps = [
+        "//prow/config:go_default_library",
         "//prow/crier:go_default_library",
         "//prow/github:go_default_library",
         "//prow/jenkins:go_default_library",
-        "//prow/jobs:go_default_library",
         "//prow/kube:go_default_library",
         "//prow/line:go_default_library",
         "//vendor:github.com/Sirupsen/logrus",
