+++ prow/plugins/BUILD
@@ -27,8 +27,8 @@ go_library(
     ],
     tags = ["automanaged"],
     deps = [
+        "//prow/config:go_default_library",
         "//prow/github:go_default_library",
-        "//prow/jobs:go_default_library",
         "//prow/kube:go_default_library",
         "//vendor:github.com/Sirupsen/logrus",
         "//vendor:github.com/ghodss/yaml",
