+++ prow/plugins/trigger/BUILD
@@ -17,9 +17,9 @@ go_test(
     library = ":go_default_library",
     tags = ["automanaged"],
     deps = [
+        "//prow/config:go_default_library",
         "//prow/github:go_default_library",
         "//prow/github/fakegithub:go_default_library",
-        "//prow/jobs:go_default_library",
         "//prow/kube:go_default_library",
         "//vendor:github.com/Sirupsen/logrus",
     ],
@@ -35,8 +35,8 @@ go_library(
     ],
     tags = ["automanaged"],
     deps = [
+        "//prow/config:go_default_library",
         "//prow/github:go_default_library",
-        "//prow/jobs:go_default_library",
         "//prow/kube:go_default_library",
         "//prow/line:go_default_library",
         "//prow/plugins:go_default_library",
