--- prow/postsubmit.yaml
@@ -1,54 +1,0 @@
----
-kubernetes/kubernetes:
-- name: ci-kubernetes-bazel-build
-  branches:
-  - master
-  spec:
-    containers:
-    - image: gcr.io/k8s-testimages/bazelbuild:0.8
-      args:
-      - "--branch=$(PULL_REFS)"
-      - "--upload=gs://kubernetes-jenkins/logs"
-      - "--git-cache=/root/.cache/git"
-      - "--clean"
-      # Bazel needs privileged mode in order to sandbox builds.
-      securityContext:
-        privileged: true
-  run_after_success:
-  - name: ci-kubernetes-bazel-test
-    spec:
-      containers:
-      - image: gcr.io/k8s-testimages/bazelbuild:0.8
-        args:
-        - "--branch=$(PULL_REFS)"
-        - "--upload=gs://kubernetes-jenkins/logs"
-        - "--git-cache=/root/.cache/git"
-        - "--clean"
-        securityContext:
-          privileged: true
-
-kubernetes/test-infra:
-- name: ci-test-infra-bazel
-  branches:
-  - master
-  spec:
-    containers:
-    - image: gcr.io/k8s-testimages/bazelbuild:0.8
-      args:
-      - "--branch=$(PULL_REFS)"
-      - "--upload=gs://kubernetes-jenkins/logs"
-      - "--git-cache=/root/.cache/git"
-      - "--clean"
-      # Bazel needs privileged mode in order to sandbox builds.
-      securityContext:
-        privileged: true
-
-- name: maintenance-ci-testgrid-config-upload
-  branches:
-  - master
-  spec:
-    containers:
-    - image: gcr.io/k8s-testimages/gcloud-in-go:0.1
-      args:
-      - "--branch=$(PULL_REFS)"
-      - "--upload=gs://kubernetes-jenkins/logs"
